+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased 
 - Only ask for ai unit consumpion consent when unpacking CM packages
 - Remap entity field type names on cm package upload
+- Add `--skip-comment-upload` when unpacking CM packages
 
 # v0.38.7
 - Add `--attachments` to get comment